Add switch picture
