+++ src/App.css
@@ -11,6 +11,8 @@
   --color-text-secondary: #282525;
   --color-text-tertiary: #c1e3ed;
   --color-text-quaternary: #fff;
+
+  --bg-image: url('./../public/images/sky.jpg');
 }
 
 [data-theme='dark'] {
@@ -23,6 +25,8 @@
   --color-text-secondary: #fff;
   --color-text-tertiary: #777;
   --color-text-quaternary: #fff;
+
+  --bg-image: url('./../public/images/sky_dark.jpg');
 }
 
 main {
@@ -32,15 +36,15 @@ main {
   left: 0px;
   right: 0px;
   overflow: auto;
+
+  background-image: var(--bg-image);
+  background-color: var(--color-canvas-quaternary);
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
 }
 
 body {
   height: 100%;
   height: 100vh;
-
-  background-image: url('./../public/images/sky.jpg');
-  background-color: var(--color-canvas-quaternary);
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: cover;
 }
